Change ehcacheVersion to 3.10.issue3302-SNAPSHOT
This commit is not to be included in any pull request, of course.

The previous commit was the merge of a DaggyFix of issue-3302
into the release/3.10 branch.  That branch looks like it's
currently on a snapshot towards 3.10.9.

I loaded this commit into IntelliJ IDEA 2024-3.2.2,
using Java 8, and ran its Gradle >
ehcache3 > ehcache3 > Tasks > publishing > publishToMavenLocal
to get a jar that includes my issue-3302 fix, to test with.